pyproject.toml
[project]
name = "ai-trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "redis>=7.4.0",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
